test app: Fix contextMenus.create paramter parsing
This commit fixes the parsing for the contextMenus.create and
contextMenus.update parameters. The fix is to prevent the fields in the
parameter object from being set when the input fields don't have a
value. The internal API implementation does not consider an empty string
field to be the same as undefined, so this caused the API to not work
properly in the test app.