spec: Begin splitting up the IDL section into a new "API" area (#23)
* spec: Add domintro class to index.bs

This class can be used to document non-normative text in the
spec output. Original code from the w3c/web-locks spec source
and h/t Joshua Bell jsbell@google.com for the pointers.

* spec: Add commented text to designate headers

* spec: Add placeholder "Usage Overview" section

At the top of the spec, add a usage section for us to fill
in with an example.

* spec: Add top-level sections, move ControlledFrame

Add a Concepts top-level section to contain more information about
high-level Controlled Frame concepts that the API will refer back to.

Add an API section that will house the different API sections.

Move the top-level element definition into the API section under
an HTML Element section and remove that from the "IDL" section that
we intend to break up into smaller parts over time.