test_app: Fix addContentScripts and insertCSS
addContentScripts() should use <all_urls> to match on. This will
ensure that whatever URL the frame is displaying will match the
provided JS from the test app.

insertCSS() assumes the provided value is CSS that can be applied
to a given page. This commit fixes the CSS syntax used and adds
!important which will override any existing CSS rules for the
body's background color.