test_app: Set up autorelease (#52)
This attempts to create a workflow that will automatically run whenever a
commit changes anything in test_app. It will bump the version number
(in the test-app-vX tag), build the app, generate an update manifest, and
publish a new release with the new swbn file and an update manifest with
all known versions.

This is based on the auto-deploy setup in https://github.com/chromeos/iwa-sink,
with some additional complexity to deal with the fact that test_app isn't the
only thing in this repo.